+++ .idea/compiler.xml
@@ -6,9 +6,12 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="Grootbrein" />
         <module name="DropBookmarks" />
+        <module name="Grootbrein" />
       </profile>
     </annotationProcessing>
+  </component>
+  <component name="EclipseCompilerSettings">
+    <option name="GENERATE_NO_WARNINGS" value="true" />
   </component>
 </project>
