+++ src/main/java/com/udemy/core/Trip.java
@@ -1,4 +1,4 @@
-package com.udemy.api;
+package com.udemy.core;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.Date;
