+++ src/main/java/com/udemy/core/UserHasCar.java
@@ -1,4 +1,4 @@
-package com.udemy.api;
+package com.udemy.core;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
