--- src/main/java/com/udemy/core/mapper/UserHasCarMapper.java
@@ -1,4 +1,0 @@
-package com.udemy.core.mapper;
-
-public class UserHasCarMapper {
-}
