--- src/main/java/com/udemy/core/mapper/UserHasProjectMapper.java
@@ -1,4 +1,0 @@
-package com.udemy.core.mapper;
-
-public class UserHasProjectMapper {
-}
