--- src/main/java/com/udemy/core/mapper/UserMapper.java
@@ -1,19 +1,0 @@
-package com.udemy.core.mapper;
-
-import com.udemy.api.User;
-import org.skife.jdbi.v2.StatementContext;
-import org.skife.jdbi.v2.tweak.ResultSetMapper;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-public class UserMapper implements ResultSetMapper<User> {
-
-        public User map(int index, ResultSet resultSet, StatementContext statementContext) throws SQLException
-        {/*
-            return new User(resultSet.getLong("ID"), resultSet.getString("FIRSTNAME"),
-                    resultSet.getString("LASTNAME"),resultSet.getString("EMAIL"),
-                    resultSet.getString("ROLE"));
-        */
-        return null;}
-}
